# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: kolog API
  description: >
    실제 서버에 적용된 springdoc/swagger 없이, IDE의 OpenAPI 뷰어로 열람하기 위한
    정적 스펙 문서. 각 엔드포인트의 `description`에 수행 로직을 간단히 적었다.
  version: "1.0.0"

servers:
  - url: /api/v1

tags:
  - name: User
  - name: Log
  - name: Chat
  - name: Emotion

paths:
  /users/signup:
    post:
      tags: [User]
      summary: 회원가입
      description: 이메일 중복 확인(중복 시 409) → BCrypt 비밀번호 해시 저장 → 액세스/리프레시 토큰 발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupRequest'
      responses:
        "201":
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignupResponse'

        "409":
          description: 이미 가입된 이메일
  /users/login:
    post:
      tags: [User]
      summary: 로그인
      description: 이메일로 유저 조회 → BCrypt 비밀번호 검증(불일치/미존재 시 401) → 토큰 발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'

        "401":
          description: 이메일 또는 비밀번호 오류
  /users/me:
    get:
      tags: [User]
      summary: 내 토큰 정보 조회
      description: >
        Authorization: Bearer {token} 헤더에서 토큰 추출 → JWT 검증 후 사용자 ID 확인 →
        유저 조회(없으면 404) → id/nickname/profileImageUrl 반환
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOi..."
      responses:
        "200":
          description: 내 프로필 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        "401":
          description: 인증이 필요하거나 액세스 토큰이 유효하지 않음
        "404":
          description: 토큰의 사용자를 찾을 수 없음

  /users/{userId}:
    get:
      tags: [User]
      summary: 사용자 조회
      description: userId(path)로 유저 조회(없으면 404) → id/nickname/profileImageUrl 반환
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        "404":
          description: 사용자를 찾을 수 없음

  /users/{userId}/profile:
    get:
      tags: [User]
      summary: 프로필 조회
      description: userId(path)로 유저 조회 → 프로필(닉네임/이미지) + 이메일 조합해 반환
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 프로필 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'

    patch:
      tags: [User]
      summary: 프로필 수정
      description: >
        닉네임 변경 시 중복 확인 후 갱신, 프로필 이미지 첨부 시 파일 저장 후 기존 파일 삭제 →
        갱신된 프로필 + 이메일 반환
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: nickname
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                profileImage:
                  type: string
                  format: binary
      responses:
        "200":
          description: 프로필 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'

  /logs:
    get:
      tags: [Log]
      summary: 날짜·시간대별 로그 목록 조회
      description: >
        date만 지정 시 hour 조건 없이 해당 날짜 전체 조회, 둘 다 생략 시 현재 날짜·시간 사용,
        hour만 지정하고 date는 생략하면 잘못된 파라미터로 400 → 결과는 date 오름차순,
        hour 오름차순으로 정렬해 반환. hour는 0~23만 허용
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            example: "2026-01-01"
        - name: hour
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 23
      responses:
        "200":
          description: 조회 성공(결과 없으면 빈 배열)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogResponse'
        "400":
          description: 잘못된 날짜, hour 범위(0~23) 초과, 또는 date 없이 hour만 지정
    post:
      tags: [Log]
      summary: 로그(영상) 생성
      description: >
        인증된 사용자 확인 → 영상 형식 검증 → 영상 저장 → 로그 생성
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                videoFile:
                  type: string
                  format: binary
                caption:
                  type: string
                date:
                  type: string
                  example: "2026-01-01"
                term:
                  type: integer
      responses:
        "201":
          description: 로그 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'

        "400":
          description: 영상 누락, 지원하지 않는 영상 형식 또는 잘못된 날짜
        "401":
          description: 인증이 필요하거나 액세스 토큰이 유효하지 않음
        "404":
          description: 로그를 작성할 사용자를 찾을 수 없음
        "500":
          description: 업로드 영상 스트림 읽기 중 서버 I/O 오류

  /logs/{logId}:
    get:
      tags: [Log]
      summary: 로그 단건 조회
      description: logId(path)로 로그 조회(없으면 404) → 작성자 정보(id/nickname/profileImageUrl)와 함께 반환
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
        "404":
          description: 로그를 찾을 수 없음
    patch:
      tags: [Log]
      summary: 로그 수정 (영상/캡션)
      description: >
        로그 조회(없으면 404) → 요청자가 작성자인지 확인(아니면 403) →
        전달된 영상과 캡션 중 있는 값만 수정. 영상 교체 시 이전 영상 파일은 커밋 후 삭제
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                caption:
                  type: string
                videoFile:
                  type: string
                  format: binary
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
        "400":
          description: caption과 videoFile 모두 없거나 지원하지 않는 영상 형식
        "401":
          description: 인증이 필요하거나 액세스 토큰이 유효하지 않음
        "403":
          description: 본인이 작성한 로그가 아님
        "404":
          description: 로그를 찾을 수 없음

  /logs/hours:
    get:
      tags: [Log]
      summary: 전체 로그 시간대 목록 조회
      description: 전체 로그를 날짜별로 그룹핑해 시간대 목록 구성
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogGetHourList'

  /logs/delete/{logId}:
    delete:
      tags: [Log]
      summary: 로그 삭제
      description: 로그 조회(없으면 404) → 요청자가 작성자인지 확인(아니면 예외) → 로그 삭제 및 영상 파일 삭제
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 삭제 성공 (본문 없음)

  /video/chat:
    post:
      tags: [Chat]
      summary: 댓글 작성
      description: 유저·로그 존재 확인 → 댓글 저장. 응답은 요청에 담긴 chatContent 문자열 그대로 에코
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCreateRequest'
      responses:
        "200":
          description: 작성 성공
          content:
            application/json:
              schema:
                type: string

  /video/{logId}/chat:
    get:
      tags: [Chat]
      summary: 댓글 목록 조회
      description: 해당 로그의 댓글 목록을 작성자 정보와 함께 조회
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatGetListResponse'

  /video/emotion:
    post:
      tags: [Emotion]
      summary: 감정표현 남기기
      description: >
        동일 유저가 해당 로그에 이미 감정표현을 남겼는지 확인(있으면 예외) →
        유저·로그 존재 확인 → 감정표현 저장. 응답은 요청에 담긴 emotionId 문자열 그대로 에코
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmotionCreateRequest'
      responses:
        "200":
          description: 등록 성공
          content:
            application/json:
              schema:
                type: string

components:
  schemas:
    UserSignupRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        nickname:
          type: string

    UserSignupResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string

    UserLoginResponse:
      type: object
      properties:
        grantType:
          type: string
          example: "Bearer"
        accessToken:
          type: string
        refreshToken:
          type: string

    UserMeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        profileImageUrl:
          type: string
          nullable: true

    UserProfileResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        profileImage:
          type: string
          nullable: true
        email:
          type: string

    LogUploaderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        profileImageUrl:
          type: string
          nullable: true

    LogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        uploader:
          $ref: '#/components/schemas/LogUploaderResponse'
        videoUrl:
          type: string
        caption:
          type: string
          nullable: true
        date:
          type: string
        hour:
          type: integer

    LogGetListResponse:
      type: object
      properties:
        date:
          type: string
        hour:
          type: integer
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        profileImage:
          type: string
          nullable: true
        videoUrl:
          type: string
        caption:
          type: string
          nullable: true

    LogGetByHourResponse:
      type: object
      properties:
        logId:
          type: integer
          format: int64
        videoUrl:
          type: string
        caption:
          type: string
          nullable: true
        date:
          type: string
        hour:
          type: integer
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        profileImage:
          type: string
          nullable: true

    LogGetByHourListResponse:
      type: object
      properties:
        hours:
          type: array
          items:
            type: integer
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogGetByHourResponse'

    LogGetHourList:
      type: object
      properties:
        date:
          type: string
        hours:
          type: array
          items:
            type: integer

    ChatCreateRequest:
      type: object
      properties:
        logId:
          type: integer
          format: int64
        chatContent:
          type: string

    ChatGetListResponse:
      type: object
      properties:
        chatId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        profileImage:
          type: string
          nullable: true
        chatContent:
          type: string

    EmotionCreateRequest:
      type: object
      properties:
        logId:
          type: integer
          format: int64
        emotionId:
          type: string